refactor(web): move VirtualDraftTable into the shared table directory
Relocate the draft table shell from the draft helpers directory to the shared table directory, alongside the read-only table shell, so both virtualized table shells live together. Pure move with no behavior change; the draft barrel and the two consumers are updated to import it from the new location.
